Update grid_df
Pretty sure this is the correct version. Simulated state spaced model with parameters: mu=0.1, sigma=0.1, delta=0.005. The brutal-force optimal point lies inside boundaries of all dimensions. Note that the simulation is seed-controlled, and the particle filter algorithm is also seed-controlled. So multiple runs with the same set of parameters yield the same results.